backend_ws/requirements.txt
fastapi
uvicorn
python-socketio
redis
motor
python-dotenv 